+++ packages/blocks-core/js/libs/wordpress/buttons/test/functionality.blocks-1.e2e.cy.js
@@ -58,6 +58,7 @@ describe('Buttons Block', () => {
 				'elements/code',
 				'elements/span',
 				'elements/mark',
+				'core/button',
 			],
 			true
 		);