+++ packages/blocks-core/js/libs/wordpress/post-navigation-link/index.js
@@ -8,7 +8,11 @@ import { __ } from '@wordpress/i18n';
 /**
  * Blockera dependencies
  */
-import { SharedBlockExtension } from '@blockera/editor';
+import {
+	SharedBlockExtension,
+	generalBlockStates,
+	sharedBlockStates,
+} from '@blockera/editor';
 import { Icon } from '@blockera/icons';
 
 /**
@@ -36,6 +40,18 @@ export const PostNavigationLink: BlockType = {
 			label: __('Link', 'blockera'),
 			description: __('Hyperlink element.', 'blockera'),
 		},
+		'elements/bold': sharedInnerBlocks['elements/bold'],
+		'elements/italic': sharedInnerBlocks['elements/italic'],
+		'elements/kbd': sharedInnerBlocks['elements/kbd'],
+		'elements/code': sharedInnerBlocks['elements/code'],
+		'elements/span': sharedInnerBlocks['elements/span'],
+		'elements/mark': sharedInnerBlocks['elements/mark'],
+	},
+	availableBlockStates: {
+		...generalBlockStates,
+		focus: generalBlockStates.focus,
+		active: sharedBlockStates.active,
+		visited: sharedBlockStates.visited,
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;