+++ packages/blocks-core/js/libs/wordpress/post-template/index.js
@@ -16,6 +16,12 @@ export const PostTemplate: BlockType = {
 	targetBlock: 'core/post-template',
 	blockeraInnerBlocks: {
 		'elements/link': sharedInnerBlocks['elements/link'],
+		'elements/bold': sharedInnerBlocks['elements/bold'],
+		'elements/italic': sharedInnerBlocks['elements/italic'],
+		'elements/kbd': sharedInnerBlocks['elements/kbd'],
+		'elements/code': sharedInnerBlocks['elements/code'],
+		'elements/span': sharedInnerBlocks['elements/span'],
+		'elements/mark': sharedInnerBlocks['elements/mark'],
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;