+++ packages/blocks-core/php/libs/wordpress/details/block.php
@@ -36,6 +36,24 @@ return array_merge(
 				'blockera/core/paragraph'      => [
 					'root' => 'p',
 				],
+				'blockera/elements/bold' => [
+					'root' => 'strong, b',
+				],
+				'blockera/elements/italic' => [
+					'root' => 'em, i',
+				],
+				'blockera/elements/kbd' => [
+					'root' => 'kbd',
+				],
+				'blockera/elements/code' => [
+					'root' => 'code',
+				],
+				'blockera/elements/span' => [
+					'root' => 'span',
+				],
+				'blockera/elements/mark' => [
+					'root' => 'mark',
+				],
 			],
 		),
 	]