+++ packages/blocks-core/php/libs/wordpress/pullquote/block.php
@@ -17,6 +17,24 @@ return array_merge(
 					'blockera/elements/citation' => [
 						'root' => 'cite',
 					],
+					'blockera/elements/bold' => [
+						'root' => 'strong, b',
+					],
+					'blockera/elements/italic' => [
+						'root' => 'em, i',
+					],
+					'blockera/elements/kbd' => [
+						'root' => 'kbd',
+					],
+					'blockera/elements/code' => [
+						'root' => 'code',
+					],
+					'blockera/elements/span' => [
+						'root' => 'span',
+					],
+					'blockera/elements/mark' => [
+						'root' => 'mark',
+					],
 				],
 				blockera_load( 'inners.link', dirname( __DIR__ ) ),
 				blockera_load( 'inners.paragraph', dirname( __DIR__ ) ),