+++ packages/wordpress/php/RenderBlock/V2/Transpiler.php
@@ -244,9 +244,8 @@ class Transpiler {
 
         // Process in a single pass.
         while ($processor->next_tag()) {
-            $id_attribute = $processor->get_attribute('id');
-            $style        = $processor->get_attribute('style');
-            $class        = $processor->get_attribute('class');
+            $style = $processor->get_attribute('style');
+            $class = $processor->get_attribute('class');
 
 			// Skip if the class contains 'be-transpiled', because it shows that the block is already transpiled.
 			if ($class && str_contains($class, 'be-transpiled')) {
@@ -272,9 +271,7 @@ class Transpiler {
 				$declarations = explode(';', $style);
 				$root_class   = str_replace('.blockera-block.', '', $selector);
 
-				if ($id_attribute) {
-					$inline_declarations[ $selector . ' #' . $id_attribute ] = $declarations;
-				} elseif (! empty(trim($class ?? '')) && ! preg_match('/wp-(block|element|elements)/i', $class) && ! str_contains($class, $root_class)) {
+				if (! empty(trim($class ?? '')) && ! preg_match('/wp-(block|element|elements)/i', $class) && ! str_contains($class, $root_class)) {
 					$inline_declarations[ $selector . ' .' . str_replace(' ', '.', $class) ] = $declarations;
 				} else {
 					$inline_declarations[ $selector ] = $declarations;